Switched from Community.Csharp.SqlLiteClient to Mono.Data.Sqlite Portable. Released Alpha 1.0
Former-commit-id: 720af33953e18d178d3deec2428656b4ddc17059

+++ Intersect Server/Classes/Entities/Player.cs
@@ -33,7 +33,7 @@ namespace Intersect_Server.Classes.Entities
 
     public class Player : Entity
     {
-        public int MyId = -1;
+        public long MyId = -1;
         public bool InGame;
         public Client MyClient;
         private bool _sentMap;
